Refactor dialog folder and remove empty CSS file
Remove the `Dialog` folder and move files to appropriate locations.

* **File Deletions:**
  - Delete `Src/Innovative.Blazor.Components/Components/Dialog/IDynamicBaseComponent.cs`
  - Delete `Src/Innovative.Blazor.Components/Components/Dialog/RightSideDialog.razor.css`
  - Delete `Src/Innovative.Blazor.Components/Components/Dialog/UIFormClass.cs`
  - Delete `Src/Innovative.Blazor.Components/Components/Dialog/UiFormField.cs`
  - Delete `Src/Innovative.Blazor.Components/Components/Dialog/UIFormViewAction.cs`

* **File Additions:**
  - Add `Src/Innovative.Blazor.Components/Components/Form/IDynamicBaseComponent.cs` with the `IDynamicBaseComponent` interface.
  - Add `Src/Innovative.Blazor.Components/Components/Form/UIFormClass.cs` with the `UIFormClass` class.
  - Add `Src/Innovative.Blazor.Components/Components/Form/UIFormField.cs` with the `UIFormFieldAttribute` class.
  - Add `Src/Innovative.Blazor.Components/Components/Form/UIFormViewAction.cs` with the `UIFormViewAction` class.

* **File Changes:**
  - Update `ExampleApp/Extensions/PasswordUpdateComponent.razor.cs` to reference the new location of `IDynamicBaseComponent`.
  - Update `Src/Innovative.Blazor.Components/Components/Detail/InnovativeDetail.razor` to reference the new location of `UIFormFieldAttribute` and `UIFormViewAction`.
  - Update `Src/Innovative.Blazor.Components/Components/Detail/InnovativeDetail.razor.cs` to reference the new location of `UIFormFieldAttribute`, `UIFormViewAction`, and `UIFormClass`.
  - Update `Src/Innovative.Blazor.Components/Components/Form/InnovativeForm.razor` to reference the new location of `UIFormFieldAttribute`.
  - Update `Src/Innovative.Blazor.Components/Components/Form/InnovativeForm.razor.cs` to reference the new location of `UIFormFieldAttribute`, `UIFormClass`, and `IDynamicBaseComponent`.
  - Update `Src/Innovative.Blazor.Components/Services/InnovativeDialogService.cs` to reference the new location of `UIFormClass`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/innovative365nl/SharedBlazorCompoments?shareId=XXXX-XXXX-XXXX-XXXX).
