+++ Src/Innovative.Blazor.Components/Components/Detail/InnovativeDetail.razor
@@ -1,5 +1,5 @@
 @using System.Reflection
-@using Innovative.Blazor.Components.Components.Dialog
+@using Innovative.Blazor.Components.Components.Form
 @using Microsoft.Extensions.Localization
 @using Innovative.Blazor.Components.Components.Text
 @using TextStyle = Innovative.Blazor.Components.Components.Text.TextStyle
