+++ Src/Innovative.Blazor.Components/Components/Form/InnovativeForm.razor
@@ -1,5 +1,5 @@
 @using System.Reflection
-@using Innovative.Blazor.Components.Components.Dialog
+@using Innovative.Blazor.Components.Components.Form
 @typeparam TModel
 
 <RadzenTemplateForm TItem="TModel" Data="@Model" Submit="@OnSubmitPressed">
